Factor out track from channel
